A lot of implmentation/fixes
I did a few stuff here, fixed all the navigation and transfer the proper information needed for each page to do their queries and stuff. Added the functionality so that you can now have messages as you return to the home page, and made messages for requested appointments and login(if you login as guest). Also added some stuff to the loading screen to make it better, added some images/logos for android use later.